Scaffolded simple Express.js app.
The Express server acts as a communication point between the mobile app and the web dashboard. The client is static and served from the "public" directory. Feel free to change up the public directory structure as you wish.

Jade view engine is included, but is only for serving up errors from the server, or other test views.